# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
...
  "/api/regular_donations":
    get:
      summary: list regular_donations
      tags:
      - Regular Donations
      security:
      - Bearer: {}
      parameters:
      - name: any_field
        in: query
        schema:
          type: string
          require: false
      - name: date
        in: query
        description: |-
          The date field accepts the yyyy/m format string for donation searches.
          For example, "2023/7" would search for donations made in July 2023.
        schema:
          type: string
          require: false
      - name: page
        in: query
        schema:
          type: string
          require: false
      - name: per_page
        in: query
        schema:
          type: string
          require: false
      - name: non_page
        in: query
        schema:
          type: string
          require: false
      responses:
        '200':
          description: successful
        '401':
          description: unauthorized
    post:
      summary: create regular_donation
      tags:
      - Regular Donations
      security:
      - Bearer: {}
      parameters: []
      responses:
        '201':
          description: Created
        '403':
          description: Current user dose not have permission
        '422':
          description: Donation_at is not future
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - home_number
            examples:
              test_donation:
                summary: Test donation create
                value:
                  home_number: TT520
                  donation_at: '2023-07-02'
                  donation_amount: 1000
                  receipt: true
                  comment: 測試用奉獻
  "/api/regular_donations/{_id}":
    parameters:
    - name: _id
      in: path
      description: _id
      required: true
      schema:
        type: string
    get:
      summary: show regular_donation
      tags:
      - Regular Donations
      security:
      - Bearer: {}
      responses:
        '200':
          description: successful
        '404':
          description: Not Found
    patch:
      summary: update regular_donation
      tags:
      - Regular Donations
      security:
      - Bearer: {}
      parameters: []
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
      requestBody:
        content:
          application/json:
            schema:
              type: object
            examples:
              test home number change:
                summary: Test regular_donation update
                value:
                  home_number: TT521
                  head_of_regular_donation_id: 2
    delete:
      summary: delete regular_donation
      tags:
      - Regular Donations
      security:
      - Bearer: {}
      responses:
        '204':
          description: successful
        '403':
          description: successful
...
  "/api/report/regular_donations":
    post:
      summary: regular_donation report
      tags:
      - Reports
      security:
      - Bearer: {}
      parameters:
      - name: test
        in: query
        schema:
          type: string
          require: false
      responses:
        '200':
          description: multi id test
        '401':
          description: Unauthorized
      requestBody:
        content:
          application/json:
            schema:
              type: object
            examples:
              report test:
                summary: Get 3 regular donations report by id
                value:
                  ids:
                  - 1
                  - 2
                  - 3
  "/api/report/special_donations":
    post:
      summary: special_donation report
      tags:
      - Reports
      security:
      - Bearer: {}
      parameters:
      - name: test
        in: query
        schema:
          type: string
          require: false
      responses:
        '400':
          description: empty id test
        '200':
          description: multi id test:false
        '401':
          description: Unauthorized
      requestBody:
        content:
          application/json:
            schema:
              type: object
            examples:
              report test:
                summary: Get 3 special donations report by id
                value:
                  ids:
                  - 1
                  - 2
                  - 3
  "/api/special_donations":
    get:
      summary: list special_donations
      tags:
      - Special Donations
      security:
      - Bearer: {}
      parameters:
      - name: any_field
        in: query
        schema:
          type: string
          require: false
      - name: date
        in: query
        description: |-
          The date field accepts the yyyy/m format string for donation searches.
          For example, "2023/7" would search for donations made in July 2023.
        schema:
          type: string
          require: false
      - name: event_id
        in: query
        schema:
          type: string
          require: false
      - name: page
        in: query
        schema:
          type: string
          require: false
      - name: per_page
        in: query
        schema:
          type: string
          require: false
      - name: non_page
        in: query
        schema:
          type: string
          require: false
      responses:
        '200':
          description: event_id and any_field query
        '401':
          description: unauthorized
    post:
      summary: create special_donation
      tags:
      - Special Donations
      security:
      - Bearer: {}
      parameters: []
      responses:
        '201':
          description: Special home number added
        '403':
          description: Current user dose not have permission
        '422':
          description: Home number is blank
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - home_number
            examples:
              test_donation:
                summary: Test donation create
                value:
                  event_id: 1
                  home_number: TT520
                  donation_at: '2023-07-02'
                  donation_amount: 1000
                  receipt: true
                  comment: 測試用奉獻
  "/api/special_donations/{_id}":
    parameters:
    - name: _id
      in: path
      description: _id
      required: true
      schema:
        type: string
    get:
      summary: show special_donation
      tags:
      - Special Donations
      security:
      - Bearer: {}
      responses:
        '200':
          description: successful
        '404':
          description: Not Found
    patch:
      summary: update special_donation
      tags:
      - Special Donations
      security:
      - Bearer: {}
      parameters: []
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
      requestBody:
        content:
          application/json:
            schema:
              type: object
            examples:
              test home number change:
                summary: Test special_donation update
                value:
                  home_number: TT521
                  head_of_special_donation_id: 2
    delete:
      summary: delete special_donation
      tags:
      - Special Donations
      security:
      - Bearer: {}
      responses:
        '204':
          description: successful
        '403':
          description: successful
...
servers:
- url: http://163.17.135.152:7465
- url: http://{defaultHost}
  variables:
    defaultHost:
      default: 0.0.0.0:3000
components:
  securitySchemes:
    Bearer:
      description: JWT key necessary to use API calls
      type: apiKey
      name: authorization
      in: header